@@ -3,7 +3,6 @@ $(document).ready(function () {
     $("#users-table").usersTable();
     $("#shifts-table").shiftsTable();
     $("#payments-table").paymentsTable();
-    $("#user-payments-table").userPaymentsTable();
     $("#stats-table").statsTable();
     $("#rookies-table").rookiesTable();
 });
@@ -62,25 +61,25 @@ $(document).ready(function () {
 
     $.fn.selectableTable = function (tableOptions) {
         // Set the options that all of them share
-        tableOptions.iDisplayLength = 10;
-        tableOptions.bAutoWidth = false;
-        tableOptions.bJQueryUI = true;
-        tableOptions.bSort = true;
-        tableOptions.bLengthChange = true;
-        tableOptions.bStateSave = true;
-        tableOptions.sPaginationType = "full_numbers";
-        tableOptions.oLanguage = { sUrl: "/dataTables.spanish.txt" }
+        tableOptions.pageLength = 10;
+        tableOptions.autoWidth = false;
+        tableOptions.jQueryUI = true;
+        tableOptions.ordering = true;
+        tableOptions.lengthChange = true;
+        tableOptions.stateSave = true;
+        tableOptions.pagingType = "full_numbers";
+        tableOptions.language = { sUrl: "/dataTables.spanish.txt" }
 
-        return this.dataTable(tableOptions);
+        return this.DataTable(tableOptions);
     }
 
     $.fn.usersTable = function () {
         var tableOptions = {
-            aoColumns:[
+            columns:[
                 nameColumn, // Nombre
                 nameColumn, // email
-                nameColumn, // phone
-                shortNameColumn, // rol
+                shortNameColumn, // phone
+                countColumn, // Credito
                 actionColumn // Action buttons
             ]
         };
@@ -91,9 +90,9 @@ $(document).ready(function () {
         var tableOptions = {
             aaSorting: [[ 6, "asc" ]],
             aoColumns:[
-                $.extend({}, shortNameColumn, {bSortable:false}), // Dia y Hora
+                $.extend({}, nameColumn, {bSortable: false, sWidth: "120px"}), // Dia y Hora
                 shortCountColumn, // Cupo
-                $.extend({}, nameColumn, {iDataSort: 6}), // Proxima
+                $.extend({}, longNameColumn, {iDataSort: 6}), // Proxima
                 shortCountColumn, // Anotados
                 countColumn, // Estado
                 actionColumn, // Action buttons
@@ -103,39 +102,49 @@ $(document).ready(function () {
         return this.selectableTable(tableOptions);
     };
 
-    $.fn.userPaymentsTable = function () {
+    $.fn.userPaymentsTable = function (options) {
         var tableOptions = {
-            aaSorting: [[ 4, "desc" ]],
-            aoColumns: [
-                $.extend({}, longNameColumn, {iDataSort: 4, mData: "created_at_formatted"}), // Fecha
-                $.extend({}, nameColumn, {mData: "month"}), // Mes
-                $.extend({}, countColumn, {mData: "amount"}), // Monto
-                $.extend({}, countColumn, {mData: "credit"}), // Credito
-                $.extend({}, hiddenColumn, {mData: "created_at"}) // Fecha sin formato
+            order: [[ 4, "desc" ]],
+            columns: [
+                $.extend({}, longNameColumn, {dataSort: 4, data: "created_at_formatted"}), // Fecha
+                $.extend({}, nameColumn, {data: "month_year_formatted"}), // Mes
+                $.extend({}, countColumn, {data: "amount"}), // Monto
+                $.extend({}, countColumn, {data: "credit"}), // Credito
+                $.extend({}, hiddenColumn, {data: "created_at"}) // Fecha sin formato
             ],
-            bProcessing: true,
-            fnRowCallback: function(nRow, aData, iDisplayIndex, iDisplayIndexFull) {
+            processing: true,
+            rowCallback: function(nRow, aData, iDisplayIndex, iDisplayIndexFull) {
                 if (aData.credit == "1") {
                     $('td', nRow).each(function() { $(this).attr('class', 'conditionalRowColor'); })
                 }
                 return nRow;
             }
         };
+        $.extend(tableOptions, options);
         return this.selectableTable(tableOptions);
     };
 
     $.fn.paymentsTable = function () {
         var tableOptions = {
-            aaSorting: [[ 6, "asc" ]],
-            aoColumns:[
-                nameColumn, // Client
-                nameColumn, // Email
-                $.extend({}, longNameColumn, {iDataSort: 6}), // Fecha
-                countColumn, // Monto
-                countColumn, // Credito
-                shortActionColumn, // Action buttons
-                hiddenColumn
-            ]
+            order: [[ 5, "asc" ]],
+            columns: [
+                $.extend({}, longNameColumn, {dataSort: 5, data: "created_at_formatted"}), // Fecha
+                $.extend({}, nameColumn, {data: "user.full_name"}), // Mes
+                $.extend({}, countColumn, {data: "amount"}), // Monto
+                $.extend({}, countColumn, {data: "credit"}), // Credito
+                $.extend({}, shortNameColumn, {dataSort: 6, data: "month_year_formatted"}), // Mes
+                $.extend({}, hiddenColumn, {data: "created_at"}), // Fecha sin formato
+                $.extend({}, hiddenColumn, {data: "month_year"}) // Mes y año sin formato
+            ],
+            processing: true,
+            footerCallback: function ( nRow, aaData, iStart, iEnd, aiDisplay ) {
+                var total = 0;
+                for (var i = 0; i < aaData.length; i++) {
+                    total += aaData[i].amount * 1;
+                }
+                var nCells = nRow.getElementsByTagName('th');
+                nCells[0].innerHTML = '<strong>Total pagado en el período: $' + total + '</strong>';
+            }
         };
         return this.selectableTable(tableOptions);
     };
