@@ -1,6 +1,7 @@
 //= require jquery
 //= require jquery_ujs
-//= require 'jquery-ui.min.js'
 //= require 'jquery-cookie/jquery.cookie.js'
 //= require 'main.js'
 //= require select2
+//= require jquery-ui
+//= require jquery-ui/datepicker-es
